airflow/requirements.txt
apache-airflow-providers-amazon
pandas
sqlalchemy
pyspark
spotipy
requests
python-dotenv
boto3